Fix asan failure
The previous asan failure is because we fail to call the dtors. This
patch fixes this by introducing a base class called `ASTNode`. All
instances of `Decl`, `Stmt`, `Type` should inheritant from it. In the
dtor of ASTContext, we'll cast the memory to ASTNode and call its dtor,
so the object could be destroyed correctly.
Signed-off-by: Jun Zhang <jun@junz.org>